Don't cover py3 compat test code
This code DOES get excuted... In fact coverage.py supports a -p option
for combining coverage from multiple runs. Ideally I could somehow use
that with tox to combine the coverage results from py2 and py3, which
would show that in fact there is still 100% code coverage. But currently
it does not seem possible with tox.

Perhaps I can make a wrapper around tox to create a named temporary
directory first, and export that path to an environment variable so
maybe it'd be available to the different tox commands.